add explicit conversion

+++ library/include/RemoteTools.h
@@ -74,7 +74,7 @@ namespace DFHack
      */
     template<class T>
     void flagarray_to_ints(RepeatedField<google::protobuf::int32> *pf, const BitArray<T> &val) {
-        for (size_t i = 0; i < val.size()*8; i++)
+        for (size_t i = 0; i < size_t(val.size()*8); i++)
             if (val.is_set(T(i)))
                 pf->Add(i);
     }
